filestore: Changes for issue #14934 - filestore related changes only

+++ src/os/filestore/FileStore.cc
@@ -1386,7 +1386,7 @@ int FileStore::mount()
 	 << cpp_strerror(ret) << dendl;
     goto close_fsid_fd;
   } else if (ret == 0) {
-    if (do_update || (int)version_stamp < g_conf->filestore_update_to) {
+    if (do_update || version_stamp < g_conf->filestore_update_to) {
       derr << "FileStore::mount: stale version stamp detected: "
 	   << version_stamp
 	   << ". Proceeding, do_update "
@@ -1758,7 +1758,7 @@ int FileStore::mount()
   timer.init();
 
   // upgrade?
-  if (g_conf->filestore_update_to >= (int)get_target_version()) {
+  if (g_conf->filestore_update_to >= get_target_version()) {
     int err = upgrade();
     if (err < 0) {
       derr << "error converting store" << dendl;
@@ -1971,7 +1971,7 @@ void FileStore::_do_op(OpSequencer *osr, ThreadPool::TPHandle &handle)
   if (g_conf->filestore_inject_stall) {
     int orig = g_conf->filestore_inject_stall;
     dout(5) << "_do_op filestore_inject_stall " << orig << ", sleeping" << dendl;
-    for (int n = 0; n < g_conf->filestore_inject_stall; n++)
+    for (uint32_t n = 0; n < g_conf->filestore_inject_stall; n++)
       sleep(1);
     g_conf->set_val("filestore_inject_stall", "0");
     dout(5) << "_do_op done stalling" << dendl;
